# Add StreamReader BaseData.Reader