--- a/Common/Data/Market/TradeBar.cs
+++ b/Common/Data/Market/TradeBar.cs
@@ -15,6 +15,7 @@
 
 using System;
 using System.Globalization;
+using System.IO;
 using System.Threading;
 using QuantConnect.Logging;
 using QuantConnect.Util;
@@ -29,7 +30,7 @@ namespace QuantConnect.Data.Market
     public class TradeBar : BaseData, IBaseDataBar
     {
         // scale factor used in QC equity/forex data files
-        private const decimal _scaleFactor = 1/10000m;
+        private const decimal _scaleFactor = 1 / 10000m;
 
         private int _initialized;
         private decimal _open;
@@ -227,7 +228,67 @@ namespace QuantConnect.Data.Market
             }
 
             // if we couldn't parse it above return a default instance
-            return new TradeBar{Symbol = config.Symbol, Period = config.Increment};
+            return new TradeBar { Symbol = config.Symbol, Period = config.Increment };
+        }
+
+        /// <summary>
+        /// TradeBar Reader: Fetch the data from the QC storage and feed it directly from the stream into the engine.
+        /// </summary>
+        /// <param name="config">Symbols, Resolution, DataType, </param>
+        /// <param name="stream">The file data stream</param>
+        /// <param name="date">Date of this reader request</param>
+        /// <param name="isLiveMode">true if we're in live mode, false for backtesting mode</param>
+        /// <returns>Enumerable iterator for returning each line of the required data.</returns>
+        public override BaseData Reader(SubscriptionDataConfig config, StreamReader stream, DateTime date, bool isLiveMode)
+        {
+            //Handle end of file:
+            if (stream == null || stream.EndOfStream)
+            {
+                return null;
+            }
+            if (isLiveMode)
+            {
+                return new TradeBar();
+            }
+
+            try
+            {
+                switch (config.SecurityType)
+                {
+                    //Equity File Data Format:
+                    case SecurityType.Equity:
+                        return ParseEquity(config, stream, date);
+
+                    //FOREX has a different data file format:
+                    case SecurityType.Forex:
+                        return ParseForex(config, stream, date);
+
+                    case SecurityType.Crypto:
+                        return ParseCrypto(config, stream, date);
+
+                    case SecurityType.Cfd:
+                        return ParseCfd(config, stream, date);
+
+                    case SecurityType.Option:
+                        return ParseOption(config, stream, date);
+
+                    case SecurityType.Future:
+                        return ParseFuture(config, stream, date);
+
+                }
+            }
+            catch (Exception err)
+            {
+                Log.Error(Invariant($"TradeBar.Reader(): Error parsing line: '{stream}', Symbol: {config.Symbol.Value}, SecurityType: ") +
+                          Invariant($"{config.SecurityType}, Resolution: {config.Resolution}, Date: {date:yyyy-MM-dd}, Message: {err}")
+                );
+            }
+
+            // we need to consume a line anyway, to advance the stream
+            stream.ReadLine();
+
+            // if we couldn't parse it above return a default instance
+            return new TradeBar { Symbol = config.Symbol, Period = config.Increment };
         }
 
         /// <summary>
@@ -273,6 +334,44 @@ namespace QuantConnect.Data.Market
             return tradeBar;
         }
 
+        /// <summary>
+        /// Parses equity trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
+        /// </summary>
+        /// <param name="config">Symbols, Resolution, DataType, </param>
+        /// <param name="streamReader">The data stream of the requested file</param>
+        /// <param name="date">Date of this reader request</param>
+        /// <returns></returns>
+        public static TradeBar ParseEquity(SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+        {
+            var tradeBar = new TradeBar
+            {
+                Symbol = config.Symbol,
+                Period = config.Increment
+            };
+            ParseEquity(tradeBar, config, streamReader, date);
+            return tradeBar;
+        }
+
+        private static void ParseEquity(TradeBar tradeBar, SubscriptionDataConfig config, StreamReader stream, DateTime date)
+        {
+            if (config.Resolution == Resolution.Daily || config.Resolution == Resolution.Hour)
+            {
+                // hourly and daily have different time format, and can use slow, robust c# parser.
+                tradeBar.Time = stream.GetDateTime().ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+            else
+            {
+                // Using custom "ToDecimal" conversion for speed on high resolution data.
+                tradeBar.Time = date.Date.AddMilliseconds(stream.GetInt32()).ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+
+            tradeBar.Open = stream.GetDecimal() * _scaleFactor;
+            tradeBar.High = stream.GetDecimal() * _scaleFactor;
+            tradeBar.Low = stream.GetDecimal() * _scaleFactor;
+            tradeBar.Close = stream.GetDecimal() * _scaleFactor;
+            tradeBar.Volume = stream.GetDecimal();
+        }
+
         private static void ParseEquity(TradeBar tradeBar, SubscriptionDataConfig config, string line, DateTime date)
         {
             var csv = line.ToCsv(6);
@@ -287,10 +386,10 @@ namespace QuantConnect.Data.Market
                 tradeBar.Time = date.Date.AddMilliseconds(csv[0].ToInt32()).ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
             }
 
-            tradeBar.Open = csv[1].ToDecimal()*_scaleFactor;
-            tradeBar.High = csv[2].ToDecimal()*_scaleFactor;
-            tradeBar.Low = csv[3].ToDecimal()*_scaleFactor;
-            tradeBar.Close = csv[4].ToDecimal()*_scaleFactor;
+            tradeBar.Open = csv[1].ToDecimal() * _scaleFactor;
+            tradeBar.High = csv[2].ToDecimal() * _scaleFactor;
+            tradeBar.Low = csv[3].ToDecimal() * _scaleFactor;
+            tradeBar.Close = csv[4].ToDecimal() * _scaleFactor;
             tradeBar.Volume = csv[5].ToDecimal();
         }
 
@@ -331,6 +430,25 @@ namespace QuantConnect.Data.Market
             ParseForex(tradeBar, config, line, date);
 
             return tradeBar;
+        }
+
+        private static void ParseForex(TradeBar tradeBar, SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+        {
+            if (config.Resolution == Resolution.Daily || config.Resolution == Resolution.Hour)
+            {
+                // hourly and daily have different time format, and can use slow, robust c# parser.
+                tradeBar.Time = streamReader.GetDateTime().ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+            else
+            {
+                //Fast decimal conversion
+                tradeBar.Time = date.Date.AddMilliseconds(streamReader.GetInt32()).ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+
+            tradeBar.Open = streamReader.GetDecimal();
+            tradeBar.High = streamReader.GetDecimal();
+            tradeBar.Low = streamReader.GetDecimal();
+            tradeBar.Close = streamReader.GetDecimal();
         }
 
         private static void ParseForex(TradeBar tradeBar, SubscriptionDataConfig config, string line, DateTime date)
@@ -391,6 +509,24 @@ namespace QuantConnect.Data.Market
             return tradeBar;
         }
 
+        /// <summary>
+        /// Parses crypto trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
+        /// </summary>
+        /// <param name="config">Symbols, Resolution, DataType, </param>
+        /// <param name="streamReader">The data stream of the requested file</param>
+        /// <param name="date">The base data used to compute the time of the bar since the line specifies a milliseconds since midnight</param>
+        public static TradeBar ParseCrypto(SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+        {
+            var tradeBar = new TradeBar
+            {
+                Symbol = config.Symbol,
+                Period = config.Increment
+            };
+            ParseCrypto(tradeBar, config, streamReader, date);
+
+            return tradeBar;
+        }
+
         private static void ParseCrypto(TradeBar tradeBar, SubscriptionDataConfig config, string line, DateTime date)
         {
             var csv = line.ToCsv(6);
@@ -412,6 +548,26 @@ namespace QuantConnect.Data.Market
             tradeBar.Volume = csv[5].ToDecimal();
         }
 
+        private static void ParseCrypto(TradeBar tradeBar, SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+        {
+            if (config.Resolution == Resolution.Daily || config.Resolution == Resolution.Hour)
+            {
+                // hourly and daily have different time format, and can use slow, robust c# parser.
+                tradeBar.Time = streamReader.GetDateTime().ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+            else
+            {
+                //Fast decimal conversion
+                tradeBar.Time = date.Date.AddMilliseconds(streamReader.GetInt32()).ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+
+            tradeBar.Open = streamReader.GetDecimal();
+            tradeBar.High = streamReader.GetDecimal();
+            tradeBar.Low = streamReader.GetDecimal();
+            tradeBar.Close = streamReader.GetDecimal();
+            tradeBar.Volume = streamReader.GetDecimal();
+        }
+
         /// <summary>
         /// Parses forex trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
         /// </summary>
@@ -427,6 +583,24 @@ namespace QuantConnect.Data.Market
                 Period = config.Increment
             };
             ParseForex(tradeBar, config, line, date);
+            return tradeBar;
+        }
+
+        /// <summary>
+        /// Parses forex trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
+        /// </summary>
+        /// <param name="config">Symbols, Resolution, DataType, </param>
+        /// <param name="streamReader">The data stream of the requested file</param>
+        /// <param name="date">The base data used to compute the time of the bar since the line specifies a milliseconds since midnight</param>
+        /// <returns></returns>
+        public static TradeBar ParseForex(SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+        {
+            var tradeBar = new TradeBar
+            {
+                Symbol = config.Symbol,
+                Period = config.Increment
+            };
+            ParseForex(tradeBar, config, streamReader, date);
             return tradeBar;
         }
 
@@ -456,6 +630,19 @@ namespace QuantConnect.Data.Market
         {
             // CFD has the same data format as Forex
             return ParseForex(config, line, date);
+        }
+
+        /// <summary>
+        /// Parses CFD trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
+        /// </summary>
+        /// <param name="config">Symbols, Resolution, DataType, </param>
+        /// <param name="streamReader">The data stream of the requested file</param>
+        /// <param name="date">The base data used to compute the time of the bar since the line specifies a milliseconds since midnight</param>
+        /// <returns></returns>
+        public static TradeBar ParseCfd(SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+        {
+            // CFD has the same data format as Forex
+            return ParseForex(config, streamReader, date);
         }
 
         /// <summary>
@@ -492,6 +679,80 @@ namespace QuantConnect.Data.Market
             tradeBar.Low = csv[3].ToDecimal() * _scaleFactor;
             tradeBar.Close = csv[4].ToDecimal() * _scaleFactor;
             tradeBar.Volume = csv[5].ToDecimal();
+
+            return tradeBar;
+        }
+
+        /// <summary>
+        /// Parses Option trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
+        /// </summary>
+        /// <typeparam name="T">The requested output type, must derive from TradeBar</typeparam>
+        /// <param name="config">Symbols, Resolution, DataType, </param>
+        /// <param name="streamReader">The data stream of the requested file</param>
+        /// <param name="date">The base data used to compute the time of the bar since the line specifies a milliseconds since midnight</param>
+        /// <returns></returns>
+        public static T ParseOption<T>(SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+            where T : TradeBar, new()
+        {
+            var tradeBar = new T
+            {
+                Period = config.Increment,
+                Symbol = config.Symbol
+            };
+
+            if (config.Resolution == Resolution.Daily || config.Resolution == Resolution.Hour)
+            {
+                // hourly and daily have different time format, and can use slow, robust c# parser.
+                tradeBar.Time = streamReader.GetDateTime().ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+            else
+            {
+                // Using custom "ToDecimal" conversion for speed on high resolution data.
+                tradeBar.Time = date.Date.AddMilliseconds(streamReader.GetInt32()).ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+
+            tradeBar.Open = streamReader.GetDecimal() * _scaleFactor;
+            tradeBar.High = streamReader.GetDecimal() * _scaleFactor;
+            tradeBar.Low = streamReader.GetDecimal() * _scaleFactor;
+            tradeBar.Close = streamReader.GetDecimal() * _scaleFactor;
+            tradeBar.Volume = streamReader.GetDecimal();
+
+            return tradeBar;
+        }
+
+        /// <summary>
+        /// Parses Future trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
+        /// </summary>
+        /// <typeparam name="T">The requested output type, must derive from TradeBar</typeparam>
+        /// <param name="config">Symbols, Resolution, DataType, </param>
+        /// <param name="streamReader">The data stream of the requested file</param>
+        /// <param name="date">The base data used to compute the time of the bar since the line specifies a milliseconds since midnight</param>
+        /// <returns></returns>
+        public static T ParseFuture<T>(SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+            where T : TradeBar, new()
+        {
+            var tradeBar = new T
+            {
+                Period = config.Increment,
+                Symbol = config.Symbol
+            };
+
+            if (config.Resolution == Resolution.Daily || config.Resolution == Resolution.Hour)
+            {
+                // hourly and daily have different time format, and can use slow, robust c# parser.
+                tradeBar.Time = streamReader.GetDateTime().ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+            else
+            {
+                // Using custom "ToDecimal" conversion for speed on high resolution data.
+                tradeBar.Time = date.Date.AddMilliseconds(streamReader.GetInt32()).ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+            }
+
+            tradeBar.Open = streamReader.GetDecimal();
+            tradeBar.High = streamReader.GetDecimal();
+            tradeBar.Low = streamReader.GetDecimal();
+            tradeBar.Close = streamReader.GetDecimal();
+            tradeBar.Volume = streamReader.GetDecimal();
 
             return tradeBar;
         }
@@ -534,7 +795,6 @@ namespace QuantConnect.Data.Market
             return tradeBar;
         }
 
-
         /// <summary>
         /// Parses Option trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
         /// </summary>
@@ -547,6 +807,17 @@ namespace QuantConnect.Data.Market
             return ParseOption<TradeBar>(config, line, date);
         }
 
+        /// <summary>
+        /// Parses Option trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
+        /// </summary>
+        /// <param name="config">Symbols, Resolution, DataType, </param>
+        /// <param name="streamReader">The data stream of the requested file</param>
+        /// <param name="date">The base data used to compute the time of the bar since the line specifies a milliseconds since midnight</param>
+        /// <returns></returns>
+        public static TradeBar ParseOption(SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+        {
+            return ParseOption<TradeBar>(config, streamReader, date);
+        }
 
         /// <summary>
         /// Parses Future trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
@@ -558,6 +829,18 @@ namespace QuantConnect.Data.Market
         public static TradeBar ParseFuture(SubscriptionDataConfig config, string line, DateTime date)
         {
             return ParseFuture<TradeBar>(config, line, date);
+        }
+
+        /// <summary>
+        /// Parses Future trade bar data into the specified tradebar type, useful for custom types with OHLCV data deriving from TradeBar
+        /// </summary>
+        /// <param name="config">Symbols, Resolution, DataType, </param>
+        /// <param name="streamReader">The data stream of the requested file</param>
+        /// <param name="date">The base data used to compute the time of the bar since the line specifies a milliseconds since midnight</param>
+        /// <returns></returns>
+        public static TradeBar ParseFuture(SubscriptionDataConfig config, StreamReader streamReader, DateTime date)
+        {
+            return ParseFuture<TradeBar>(config, streamReader, date);
         }
 
         /// <summary>
@@ -617,7 +900,7 @@ namespace QuantConnect.Data.Market
             if (fillForward)
             {
                 // zero volume out, since it would skew calculations in volume-based indicators
-                ((TradeBar) clone).Volume = 0;
+                ((TradeBar)clone).Volume = 0;
             }
 
             return clone;