--- a/Common/QuantConnect.csproj
+++ b/Common/QuantConnect.csproj
@@ -792,6 +792,7 @@
     <Compile Include="Util\SecurityIdentifierJsonConverter.cs" />
     <Compile Include="Util\SeriesJsonConverter.cs" />
     <Compile Include="Util\StreamReaderEnumerable.cs" />
+    <Compile Include="Util\StreamReaderExtensions.cs" />
     <Compile Include="Util\TypeChangeJsonConverter.cs" />
     <Compile Include="Util\LinqExtensions.cs" />
     <Compile Include="Util\MemoizingEnumerable.cs" />