… Common/Data/BaseData.cs
… Common/Data/IBaseData.cs
… Engine/DataFeeds/TextSubscriptionDataSourceReader.cs
… Engine/DataFeeds/Transport/LocalFileSubscriptionStreamReader.cs
… Tests/Common/Util/StreamReaderExtensionsTests.cs