<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Add `BaseData.Reader` implementation which consumes the `StreamReader`
directly, avoiding in-between substrings and parsing improving
performance and reducing resource consumption
- Adding unit tests, including performance unit test showing a >50%
improvement

`LOCAL PR EmptyMinute400EquityBenchmark - only SPY`
18.10 seconds at 235k data points per second. Processing total of 4,255,822 data points.
18.00 seconds at 236k data points per second. Processing total of 4,255,822 data points.
18.16 seconds at 234k data points per second. Processing total of 4,255,822 data points.
`LOCAL MASTER EmptyMinute400EquityBenchmark - only SPY`
19.91 seconds at 214k data points per second. Processing total of 4,255,822 data points.
20.05 seconds at 212k data points per second. Processing total of 4,255,822 data points.
19.71 seconds at 216k data points per second. Processing total of 4,255,822 data points.

`MASTER CLOUD EmptyMinute400EquityBenchmark - only SPY`
28.70 seconds at 148k data points per second. Processing total of 4,255,699 data points.
29.66 seconds at 143k data points per second. Processing total of 4,255,699 data points.
27.69 seconds at 154k data points per second. Processing total of 4,255,699 data points.
`PR Cloud EmptyMinute400EquityBenchmark - only SPY`
24.14 seconds at 176k data points per second. Processing total of 4,255,699 data points.
24.61 seconds at 173k data points per second. Processing total of 4,255,699 data points.
21.95 seconds at 194k data points per second. Processing total of 4,255,699 data points.

`PR`
![image](https://user-images.githubusercontent.com/18473240/69093020-68b8c600-0a2c-11ea-8313-811891514406.png)
`MASTER`
![image](https://user-images.githubusercontent.com/18473240/69093045-6fdfd400-0a2c-11ea-8abc-66fa2911c012.png)

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Partly implements https://github.com/QuantConnect/Lean/issues/3850, just for `TradeBars` and `TextSubscriptionDataSourceReader`
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Never fast enough
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit, regression tests, cloud testing
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x]  All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->